move elapsedTime function to time.js
